// src/components/form/types.ts
export type AccountType = 'PF' | 'PJ';

export interface FormValues {
  accountType: AccountType;
  fullName?: string;
  cpf?: string;
  birthDate?: string;
  rg?: string;
  cnh?: string;
  isForeigner?: boolean;
  userEmail?: string;
  userPhone?: string;
  pepStatus?: boolean;
  pepPosition?: string;
  cep?: string;
  street?: string;
  number?: string;
  complement?: string;
  district?: string;
  city?: string;
  state?: string;
  foreignStreet?: string;
  foreignNumber?: string;
  foreignComplement?: string;
  foreignDistrict?: string;
  foreignCity?: string;
  foreignState?: string;
  foreignZipCode?: string;
  foreignCountry?: string;
  documentFront?: File | null;
  documentBack?: File | null;
  selfie?: File | null;
  proofOfAddress?: File | null;
  companyName?: string;
  tradeName?: string;
  cnpj?: string;
  foundationDate?: string;
  mainCNAE?: string;
  companyEmail?: string;
  companyPhone?: string;
  legalNature?: string;
  pjCep?: string;
  pjStreet?: string;
  pjNumber?: string;
  pjComplement?: string;
  pjDistrict?: string;
  pjCity?: string;
  pjState?: string;
  majorityAdminName?: string;
  majorityAdminCpf?: string;
  majorityAdminEmail?: string;
  majorityAdminPhone?: string;
  articlesOfAssociation?: File | null;
  cnpjCard?: File | null;
  adminIdFront?: File | null;
  adminIdBack?: File | null;
  companyProofOfAddress?: File | null;
  ecnpjCertificate?: File | null;
  acceptTerms?: boolean;
}